+++ packages/other_packages/shared_preferences_dtx/lib/src/shared_preferences_extensions.dart
@@ -1,3 +1,7 @@
+import 'dart:convert';
+import 'dart:ui';
+
+import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 ///
@@ -9,8 +13,10 @@ extension SharedPreferencesExt on SharedPreferences {
   /// Reads a value, throwing an exception if it's not a
   /// double, int, string, bool or List<String>.
   T? getValue<T>(String key) {
-    assert(T != dynamic,
-        'Incompatible dynamic type - Use only double, int, String, bool or List<String>');
+    assert(
+        T != dynamic,
+        'Incompatible dynamic type - Use only double, int, String, bool, List<String>, '
+        'Map<String, dynamic>, ThemeMode or Color');
     if (T == double) {
       return getDouble(key) as T?;
     } else if (T == int) {
@@ -21,9 +27,16 @@ extension SharedPreferencesExt on SharedPreferences {
       return getStringList(key) as T?;
     } else if (T == bool) {
       return getBool(key) as T?;
+    } else if (T == Color) {
+      return getColor(key) as T?;
+    } else if (T == _typeOf<Map<String, dynamic>>()) {
+      return getJson(key) as T?;
+    } else if (T == ThemeMode) {
+      return getThemeMode(key) as T?;
     }
     throw ArgumentError(
-        'Incompatible preference type - Use only double, int, String, bool or List<String>');
+        'Incompatible preference type - Use only double, int, String, bool, List<String>, '
+        'Map<String, dynamic>, ThemeMode or Color');
   }
 
   /// Reads a value, throwing an exception if it's not a
@@ -31,8 +44,10 @@ extension SharedPreferencesExt on SharedPreferences {
   ///
   /// If the value is null, return the [defaultValue]
   T getOrElse<T>(String key, {required T defaultValue}) {
-    assert(T != dynamic,
-        'Incompatible dynamic type - Use only double, int, String, bool or List<String>');
+    assert(
+        T != dynamic,
+        'Incompatible dynamic type - Use only double, int, String, bool, List<String>, '
+        'Map<String, dynamic>, ThemeMode or Color');
     if (T == double) {
       return getDouble(key) as T? ?? defaultValue;
     } else if (T == int) {
@@ -43,9 +58,16 @@ extension SharedPreferencesExt on SharedPreferences {
       return getStringList(key) as T? ?? defaultValue;
     } else if (T == bool) {
       return getBool(key) as T? ?? defaultValue;
+    } else if (T == Color) {
+      return getColor(key) as T? ?? defaultValue;
+    } else if (T == _typeOf<Map<String, dynamic>>()) {
+      return getJson(key) as T? ?? defaultValue;
+    } else if (T == ThemeMode) {
+      return getThemeMode(key) as T? ?? defaultValue;
     }
     throw ArgumentError(
-        'Incompatible preference type - Use only double, int, String, bool or List<String>');
+        'Incompatible preference type - Use only double, int, String, bool, List<String>, '
+        'Map<String, dynamic>, ThemeMode or Color');
   }
 
   /// Writes a value, throwing an exception if it's not a
@@ -61,9 +83,39 @@ extension SharedPreferencesExt on SharedPreferences {
       return setString(key, value);
     } else if (value is bool) {
       return setBool(key, value);
+    } else if (value is Color) {
+      return setColor(key, value);
+    } else if (value is Map<String, dynamic>) {
+      return setJson(key, value);
+    } else if (value is ThemeMode) {
+      return setThemeMode(key, value);
     }
 
     throw ArgumentError(
-        'Incompatible preference type - Use only double, int, String, bool or List<String>');
+        'Incompatible preference type - Use only double, int, String, bool, List<String>, '
+        'Map<String, dynamic>, ThemeMode or Color');
+  }
+
+  Future<bool> setColor(String key, Color color) => setInt(key, color.value);
+
+  Future<bool> setThemeMode(String key, ThemeMode themeMode) =>
+      setInt(key, ThemeMode.values.indexOf(themeMode));
+
+  Future<bool> setJson(String key, Map<String, dynamic> map) =>
+      setString(key, jsonEncode(map));
+
+  Color? getColor(String key) {
+    final value = getInt(key);
+    return value != null ? Color(value) : null;
+  }
+
+  ThemeMode? getThemeMode(String key) {
+    final value = getInt(key);
+    return value != null ? ThemeMode.values[value] : null;
+  }
+
+  Map<String, dynamic>? getJson(String key) {
+    final value = getString(key);
+    return value != null ? jsonDecode(value) as Map<String, dynamic> : null;
   }
 }