+++ packages/other_packages/shared_preferences_dtx/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.0.0] - 16/02/2021
 
 - Null Safety
+- Add new preferences: Color, Map<String, dynamic> and ThemeMode.
 ## [1.0.0] - 16/12/2020
 
 Simplified way to obtain a preference `preference.getValue` instead of `preference.getString` and set a preference value `preference.setValue` instead of `preference.setString`